enhance(scripts): add support to build a package from a local source directory set in `$TERMUX_PKG_SRCURL` if its in the format `file:///path/to/source/dir`
The `/path/to/source/dir` must be an absolute and normalized path to a source directory on the local file system.

Whatever was the current state of the local source directory when build was started, that is what will be built, without any changes to any `git` branches/tags. Any uncommitted changes to current `git` branch will also get built. Any value for `TERMUX_PKG_GIT_BRANCH` in the `build.sh` of the package will be ignored since a `tar` file will be created from the source directory and it will be used as is. No checksum checks like one set in `$TERMUX_PKG_SHA256` will be done, and a tar for source directory will be created every time package is built, assuming `-f/-F` flags are passed for rebuilds, and the `-r` flag (to be added later) will not be required.

Note that to build from a local git repo, the `git+file:///path/to/source` format must be used with the `git+` prefix.